# Reduce input lag by 36ms by separating frame rate limiting from update batching