@metalefty 2020-06-11
Hi, thank you very much for your contribution. It looks good to me at first glance, let me some more time to test.