@trishume 2020-06-11
Cool no problem. I've been using it for a day and I've checked that it works in gnome-terminal, xterm, emacs and Chrome. I checked using my latency tester that the latency does indeed improve. I also verified with a high speed camera on testufo.com that the frame limit was still working, although the first time I did that test I realized I got a comparison flipped and had to fix that, hence the force push.