@metalefty 2020-06-12
Fortunately, perhaps unfortunately, I've never seen 120ms delay so I don't see the latency differences in your fix. I might be insensitive but 120ms is noticeable to me. However, your suggestion is quire right. I'm fine with this. Also, replacing hardcoded magic numbers with macros is very good, thanks.

@jsorg71 Any concerns?