@trishume 2020-06-12
Yah the 36ms improvement is only noticeable on its own if you feel for it and compare side by side, but I confirmed it's there with my latency tester, and I think these kinds of differences do subtly affect feel and they add up. If I use my Dell monitor with better latency alongside this patch I get a 30ms difference from the monitor and a 36ms difference from the patch and suddenly the latency going from 120ms to ~60ms is something I can notice and feels great.

This was just an easy thing I noticed as a prelude to trying to improve 4k performance with apps like Chrome, although I'm less sure you'll want to merge my other patches, at least without some cleanup. So far I've improved scrolling and latency in apps like Sublime Text that submit large change rectangles by switching the hashing from CRC to wyhash, and only doing the color conversion if a tile differs: https://github.com/trishume/xorgxrdp/tree/thume-remix

Haven't got around to writing up proper measurements on those changes yet though, but in the worst case I measured on a 4k monitor the capture time went from 120ms to 11ms. The wyhash header also doesn't conform to the old version of C that xorgxrdp uses.

I'm also thinking of trying to parallelize some things with OpenMP or something.