@metalefty 2020-06-12
xorgxrdp has much room for performance improvement. It has a change get merged but as you say some cleanup will be needed.