@bolkedebruin 2020-06-14
@trishume Can I say that this is awesome? We are using vnc now as backend due to the latency xorg is giving us, which is strange of course due the the "nativeness" of xorg. This latency is *very* noticable. We can compare different systems and this is currently the order that users perceive as the fastest: Citrix, RDP (W10), XRDP (vnc), XRDP (xorg), so any improvement in this area is very welcome.