@trishume 2020-06-15
@bolkedebruin Interesting. Well this patch plus the patch I linked in my other comment I haven't submitted yet should hopefully make it similar to or better than the VNC backend. I don't think I can compete with W10 RDP and Citrix yet though. That'd require some parallelizing (more annoying than I hoped) and I'm thinking even that probably won't be enough.

I use a 4k display and the absolute worst thing is scrolling, because there's just so many changed pixels so frequently. I know Citrix has some kind of scrolling detection optimization that allows it to compress by re-using a shifted version of the previous screen area. I think it *might* be possible to do this in xrdp by detecting areas that are a constant pixel offset in the frame then using `xrdp_orders_screen_blt` on that, but I haven't yet figured out if that handles overlapping regions properly.