@bolkedebruin 2020-06-15
@trishume I havent tried your alternative CRC implementation yet if that is what you are referring too. If it reduces (perceived) latency we will do that soon.

Indeed scrolling is the worst so if that could be improved I am all ears ;-). We also see that cpu usage of xrdp goes up significantly when scrolling. It can max out a cpu quite easily. That probably makes sense given all the screen changes, but nevertheless it was surprising to us. We had to make a cpu core available to xrdp per session to allow enough performance for the users.

I hop you continue providing patches in this area. I can offer a significant testing base to which we can roll out a new version pretty quickly.