@trishume 2020-06-23
@bolkedebruin so I implemented scrolling detection and it makes a big difference to scrolling. This branch has my patches with my USDT probes cut out including this PR, accelerating large screen updates by switching to wyhash, and scrolling detection: https://github.com/trishume/xorgxrdp/tree/new-patch

Note that scrolling detection has unavoidable tearing at the fringes between where it does the scroll rectangle and the updated tiles. It can also copy small tile compression artifacts making slightly noticeable color fringes. Both of these are only fixable by moving xrdp to use the RemoteFX Progressive protocol, see for instance https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-rdpegfx/0b19d058-fff0-43e5-8671-8c4186d60529 and https://docs.microsoft.com/en-us/openspecs/windows_protocols/ms-rdpegfx/f6cbf151-d83e-4f7a-a290-809e376f7722

In order to get the scrolling detection done and have fun doing it in my spare time I completely ignored the xorgxrdp code style. So the patch includes long lines and more modern C where I declare variables where I use them. I also allocate some small temporary buffers during the scroll detection. All of these are things I'm guessing would be desirable to fix if it was to be upstreamed. Fixing the code style of wyhash might be especially annoying if that's required. It could also use some more comments explaining how it works, although that would be very against the xorgxrdp code style 😛 

In the mean time people who want dramatically better performance in common situations but don't care about compatibility with super old C versions (i.e. basically everyone) can use my branch. Hopefully I or someone else will bother to clean up my patch for upstreaming at some point.