@trishume 2020-06-24
@bolkedebruin note that I've submitted these changes as a work in progress PR so I guess further comments should go on https://github.com/neutrinolabs/xorgxrdp/pull/167